--- 03-frontend-masters/11-intro-to-realtime/polling/backoff-and-retry/frontend/style.css
@@ -1,13 +1,0 @@
-.page-column {
-  max-width: 600px;
-  margin: 30px auto;
-}
-
-h1 {
-  text-align: center;
-  color: #26a69a;
-}
-
-.btn-box {
-  text-align: center;
-}
