+++ 03-frontend-masters/11-intro-to-realtime/polling/exercise/backend/server.js
@@ -21,16 +21,21 @@ app.use(bodyParser.json());
 app.use(express.static("frontend"));
 
 app.get("/poll", function (req, res) {
-  // use getMsgs to get messages to send back
-  // write code here
+  // send 500 sometimes to check retries
+  res.status(Math.random() > 0.5 ? 200 : 500).json({
+    msg: getMsgs(),
+  });
 });
 
 app.post("/poll", function (req, res) {
-  // add a new message to the server
-  // write code here
+  const { user, text } = req.body;
+
+  msg.push({ user, text, time: Date.now() });
+
+  res.json({ status: "ok" });
 });
 
 // start the server
-const port = process.env.PORT || 3000;
+const port = process.env.PORT || 3001;
 app.listen(port);
 console.log(`listening on http://localhost:${port}`);
