+++ 03-frontend-masters/11-intro-to-realtime/polling/exercise/frontend/polling-chat.js
@@ -15,13 +15,36 @@ chat.addEventListener("submit", function (e) {
 });
 
 async function postNewMsg(user, text) {
-  // post to /poll a new message
-  // write code here
+  const data = { user, text };
+  const options = {
+    method: "POST",
+    body: JSON.stringify(data),
+    headers: { "Content-Type": "application/json" },
+  };
+
+  await fetch("/poll", options);
 }
 
 async function getNewMsgs() {
-  // poll the server
-  // write code here
+  let json;
+  try {
+    const res = await fetch("/poll");
+    json = await res.json();
+
+    if (res.status >= 400) {
+      throw new Error("Request did not succeed: " + res.status);
+    }
+
+    allChat = json.msg;
+    render();
+
+    failedTries = 0;
+  } catch (err) {
+    console.error(err);
+    failedTries++;
+  }
+
+  // setTimeout(getNewMsgs, INTERVAL);
 }
 
 function render() {
@@ -38,4 +61,24 @@ const template = (user, msg) =>
   `<li class="collection-item"><span class="badge">${user}</span>${msg}</li>`;
 
 // make the first request
-getNewMsgs();
+
+// - polling
+// getNewMsgs();
+
+//
+
+const BACKOFF = 5000;
+let timeToMakeNextRequest = 0;
+let failedTries = 0;
+
+// - pause polling when the window is unfocused.
+async function requestAnimationFrameTimer(time) {
+  if (timeToMakeNextRequest <= time) {
+    await getNewMsgs();
+    timeToMakeNextRequest = time + INTERVAL + failedTries * BACKOFF;
+  }
+
+  requestAnimationFrame(requestAnimationFrameTimer);
+}
+
+requestAnimationFrame(requestAnimationFrameTimer);
