--- 03-frontend-masters/11-intro-to-realtime/polling/pause-on-unfocus/backend/server.js
@@ -1,42 +1,0 @@
-import express from "express";
-import bodyParser from "body-parser";
-import nanobuffer from "nanobuffer";
-import morgan from "morgan";
-
-const app = express();
-const msg = new nanobuffer(50);
-const getMsgs = () => Array.from(msg).reverse();
-
-msg.push({
-  user: "brian",
-  text: "hi",
-  time: Date.now(),
-});
-
-app.use(morgan("dev"));
-app.use(bodyParser.json());
-app.use(express.static("frontend"));
-
-app.get("/poll", function (req, res) {
-  res.json({
-    msg: getMsgs(),
-  });
-});
-
-app.post("/poll", function (req, res) {
-  const { user, text } = req.body;
-
-  msg.push({
-    user,
-    text,
-    time: Date.now(),
-  });
-
-  res.json({
-    status: "ok",
-  });
-});
-
-const port = process.env.PORT || 3000;
-app.listen(port);
-console.log(`listening on http://localhost:${port}`);
